pyproject.toml
[project]
name = "rag-aws-agent"
version = "0.1.0"
description = "AI Agent RAG Application using Google ADK patterns with AWS Bedrock, S3 Vector Storage, and DynamoDB"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Developer", email = "dev@example.com"}
]
keywords = ["ai", "agent", "rag", "aws", "bedrock", "dynamodb", "s3"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "google-generativeai>=0.8.0",
    "langchain>=0.3.0",
    "langchain-aws>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0,<2.0.0",  # Pin to 1.x for pre-built wheels in Lambda
    "tiktoken>=0.7.0,<0.9.0",  # Pin for pre-built Lambda-compatible wheels
    "aiohttp>=3.9.0",
    "fastapi>=0.110.0",
    "python-multipart>=0.0.9",  # Required for FastAPI file uploads
    "uvicorn>=0.27.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "boto3-stubs[bedrock,bedrock-runtime,dynamodb,s3]>=1.34.0",
    "pre-commit>=3.6.0",
]
deploy = [
    "aws-cdk-lib>=2.120.0",
    "constructs>=10.3.0",
]

[project.scripts]
rag-agent = "rag_agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_agent"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src/rag_agent --cov-report=term-missing"

